Create DC motor speed control with a distance sensor
the complete instructions so that you don't have to jump back and forth between the two articles.
The new connections that relate to the ultrasonic distance sensor are #10, #11, #12 and #13.

Start by unplugging the Arduino from your computer so that it is not in operation.

Connect the first motor to motor controller module Out1 and Out2. The order does not matter.
Connect the first motor to motor controller module Out3 and Out4. The order does not matter.
Connect the positive wire from the battery pack to pin +12V on the module.
Connect the negative wire from the battery pack to pin GND on the module.
Connect the GND pin of the module to the GND pin of the Arduino.
Connect Arduino pin 5 to module pin In1.
Connect Arduino pin 4 to module pin In2.
Connect Arduino pin 3 to module pin In3.
Connect Arduino pin 2 to module pin In4.
Connect the sensor pin Vcc to Arduino pin 5V.
Connect the sensor pin GND to Arduino pin GND.
Connect the sensor pin Trig to Arduino pin 13.
Connect the sensor pin Echo to Arduino pin 12.
Inspect the wiring and ensure they are correct.