Code and brief understanding of PID control in Arduino by me
PID controller coded in Arduino. Now we need to modify it for the distance sensor as a setpoint and as feedback.
First we start off by defining the variables needed for the computation. And you’ll notice most of these are of the double data type or floating-point. This is needed because it allows more precise computation. Now just to make things easier I kept the computation of the control signal in a separate function called the control. Now in order to make sure that we get a uniform interval of sample time t we first start off by storing the current time the program has been running by using a built-in function in Arduino IDE called millies. This function essentially gives you the number of milliseconds passed since the start of the program. we can keep track of the current time we can compute the delta time by subtracting the current time from the previous time now to get the updated control signal every t seconds.

we have an IF condition whereby if the time interval during the loop equals or exceeds t the control signal is computed and updated. This if condition is what allows the controller to get called FBT seconds and it makes realizing the controller a lot easier. Our error term is simply the setpoint minus the sensed output the total error refers to the integral term of accumulating the error. Every t seconds the Delta error refers to the derivative term whereas just a current error minus the last error. So when those values are computed and obtained we can sum them up to form our control signal. And you can see here that the summation is pretty much the same as shown in the discrete expression. So what are the control signals computed? We actually store the error as the last error for the next iteration. And the same goes for the time. This pretty much covers how the control algorithm is implemented or on the order.
https://microcontrollerslab.com/pid-controller-implementation-using-arduino/