Motor Control System Design Code
Motor Control System Design Code  with the implementation of PID algorithm. It generates the best, stable control system with PID gains gains tuned for optimum outcome. This code is a proprietary property of Shankhaneel Basak. Any use of this script should include proper citation of the original author as per copyright policy in MATLAB 2020b